api/requirements.txt
openai==2.7.2
langchain==1.0.5
langchain-openai==1.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytest==7.4.3
rank-bm25==0.2.2
redis==5.0.1
prometheus-client==0.19.0
fastapi==0.115.6
uvicorn==0.32.1
